refactor handlers to use with repository
