+++ product-service/src/handlers/getProductById.ts
@@ -1,17 +1,17 @@
 import { APIGatewayProxyResult } from "aws-lambda";
-import { ProductServiceInterface, ProductByIdEvent, errorResponse, successResponse } from "../service";
+import { ProductByIdEvent, errorResponse, successResponse, ProductService } from "../services/product-service";
 
-export const getSingleProduct = (productService: ProductServiceInterface) => 
+export const getSingleProduct = (productService: ProductService) => 
   async (event: ProductByIdEvent): Promise<APIGatewayProxyResult> => {
     try {
       const { productId } = event.pathParameters;
       const product = await productService.getProductById(productId);
-
+      
       if (!product) {
         return errorResponse(new Error(`Product with id ${productId} not found`), 404);
       }
 
-      return successResponse({ product });
+      return successResponse(product);
     }
     catch (err: any) {
       return errorResponse(err);
